+++ lib/node/node.js
@@ -6,6 +6,8 @@ var id = require('../util/id');
 
 var QueryBuilder = require('../filter/query-builder');
 
+var Validator = require('./validator');
+
 var TYPE = require('./type');
 module.exports.TYPE = TYPE;
 var GEOMETRY = require('./geometry');
@@ -200,123 +202,38 @@ Node.prototype.toDot = function() {
     };
 };
 
-
-var TYPE = {
-    NODE: 'node',
-    NUMBER: 'number',
-    STRING: 'string',
-    BOOLEAN: 'boolean',
-    ENUM: 'enum',
-    ARRAY: 'array'
-};
-
-var GEOMETRY = {
-    ANY: '*',
-    POINT: 'point',
-    LINESTRING: 'linestring',
-    POLYGON: 'polygon'
-};
-
-module.exports.GEOMETRY = GEOMETRY;
-
-function Validator(type, isValid, nullable, expects, defaultValue) {
-    this.type = type;
-    this.isValid = isValid;
-    this.nullable = nullable || false;
-    this.expects = expects || type;
-    this.defaultValue = defaultValue || null;
-}
-
-Validator.prototype.toJSON = function() {
-    var json = {
-        type: this.type
-    };
-    if (this.nullable) {
-        json.optional = true;
-    }
-    return json;
-};
-
-Validator.prototype.getDefaultValue = function() {
-    return this.defaultValue;
-};
-
-function EnumValidator(validValues) {
-    this.type = TYPE.ENUM;
-    this.validValues = validValues;
-    this.isValid = function(val) {
-        return validValues.indexOf(val) > -1;
-    };
-    this.expects = TYPE.ENUM + '(' + validValues.map(JSON.stringify).join(',') + ')';
-}
-
-EnumValidator.prototype.toJSON = function() {
-    return {
-        type: this.type,
-        values: this.validValues
-    };
-};
-
-EnumValidator.prototype.getDefaultValue = function() {
-    return this.validValues[0];
-};
-
-function NodeValidator(geometry) {
-    this.type = TYPE.NODE;
-    this.geometry = geometry || [GEOMETRY.ANY];
-    this.geometries = this.geometry.reduce(function(geometries, geometry) {
-        geometries[geometry] = true;
-        return geometries;
-    }, {});
-    this.isValid = function (node) {
-        return !!node.getQuery;
-    };
-    this.defaultValue = null;
-}
-
-NodeValidator.prototype.toJSON = function() {
-    return {
-        type: this.type,
-        geometry: this.geometry
-    };
-};
-
-NodeValidator.prototype.getDefaultValue = function() {
-    return this.defaultValue;
-};
-
 module.exports.PARAM = {
     NODE: function() {
         var args = (arguments.length === 1 ? [arguments[0]] : Array.apply(null, arguments));
-        return new NodeValidator(args);
+        return new Validator.Node(args);
     },
     NUMBER: function() {
-        return new Validator(TYPE.NUMBER, Number.isFinite);
+        return new Validator.Generic(TYPE.NUMBER, Number.isFinite);
     },
     STRING: function() {
-        return new Validator(TYPE.STRING, function (str) {
+        return new Validator.Generic(TYPE.STRING, function (str) {
             return typeof str === 'string';
         });
     },
     BOOLEAN: function() {
-        return new Validator(TYPE.BOOLEAN, function (bool) {
+        return new Validator.Generic(TYPE.BOOLEAN, function (bool) {
             return typeof bool === 'boolean';
         });
     },
     ENUM: function() {
         var args = (arguments.length === 1 ? [arguments[0]] : Array.apply(null, arguments));
-        return new EnumValidator(args);
+        return new Validator.Enum(args);
     },
     ARRAY: function (validator) {
         if (validator) {
-            return new Validator(TYPE.ARRAY, function(arr) {
+            return new Validator.Generic(TYPE.ARRAY, function(arr) {
                 return Array.isArray(arr) && arr.every(validator.isValid);
             }, false, TYPE.ARRAY + '<' + validator.type + '>');
         }
-        return new Validator(TYPE.ARRAY, Array.isArray);
+        return new Validator.Generic(TYPE.ARRAY, Array.isArray);
     },
     NULLABLE: function (validator) {
-        return new Validator(validator.type, function(param) {
+        return new Validator.Generic(validator.type, function(param) {
             return (param === null || typeof param === 'undefined') || validator.isValid(param);
         }, true, validator.expects, validator.getDefaultValue());
     }
