+++ lib/node/validator/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+    Enum: require('./enum'),
+    Generic: require('./generic'),
+    Node: require('./node')
+};
